feat: table reading
Can work with basic table data now, provided the types of the columns
are known at compile time. Next would be to have dynamically typed
columns for use with the CLI, but that is beyond the current need so to
be delayed until later.

Split the code up into a number of smaller, more logical units.