fix(cfitsio): add wasm32 preprocessor switches
CFITSIO has fancy byte-swapping which really cares about endian-ness,
since the FITS specification states that all FITS files are big endian.
Since WASM is little endian, `BYTESWAPPED` is set to true, and we
enforce `LONGSIZE 32` for WASM32.

The switch for `unistd.h` may not work on mac or windows, depending on
libc.